+++ packages/fxa-content-server/tests/functional/sync_v2_reset_password.js
@@ -72,6 +72,8 @@ define([
         .then(closeCurrentWindow())
 
         .then(testSuccessWasShown())
+        // Only expect the login message in the verification tab to avoid
+        // a race condition within the browser when it receives two login messages.
         .then(noSuchBrowserNotification('fxaccounts:login'));
     },