+++ packages/fxa-content-server/tests/server/amplitude.js
@@ -7,25 +7,21 @@
 define([
   'intern!object',
   'intern/chai!assert',
-  'intern/dojo/node!path',
-  'intern/dojo/node!proxyquire',
   'intern/dojo/node!sinon',
+  'intern/dojo/node!../../server/lib/amplitude',
   'intern/dojo/node!../../package.json',
-], (registerSuite, assert, path, proxyquire, sinon, package) => {
+], (registerSuite, assert, sinon, amplitude, package) => {
   const APP_VERSION = /^[0-9]+\.([0-9]+)\./.exec(package.version)[1];
-
-  let log, amplitude;
 
   registerSuite({
     name: 'amplitude',
 
     beforeEach () {
-      log = {
-        info: sinon.spy()
-      };
-      amplitude = proxyquire(path.resolve('server/lib/amplitude'), {
-        './logging/log': () => log
-      });
+      sinon.stub(process.stderr, 'write', () => {});
+    },
+
+    afterEach () {
+      process.stderr.write.restore();
     },
 
     'app version seems sane': () => {
@@ -62,10 +58,12 @@ define([
         utm_source: 'bnag',
         utm_term: 'plin'
       });
-      assert.equal(log.info.callCount, 1);
-      const args = log.info.args[0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const args = process.stderr.write.args[0];
       assert.lengthOf(args, 1);
-      assert.deepEqual(args[0], {
+      assert.typeOf(args[0], 'string');
+      assert.equal(args[0][args[0].length - 1], '\n');
+      assert.deepEqual(JSON.parse(args[0]), {
         app_version: APP_VERSION,
         device_id: 'bar',
         event_properties: {
@@ -109,8 +107,8 @@ define([
         utm_source: 'l',
         utm_term: 'm'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.deepEqual(arg, {
         app_version: APP_VERSION,
         device_id: 'b',
@@ -142,8 +140,8 @@ define([
         lang: 'd',
         uid: 'none'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_pref - disconnect_device');
       assert.isUndefined(arg.device_id);
       assert.isUndefined(arg.event_properties.device_id);
@@ -160,8 +158,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_pref - logout');
     },
 
@@ -174,8 +172,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - engage');
     },
 
@@ -188,8 +186,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - engage');
     },
 
@@ -211,8 +209,8 @@ define([
         utm_source: 'l',
         utm_term: 'm'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_reg - engage');
       assert.deepEqual(arg, {
         app_version: APP_VERSION,
@@ -249,7 +247,7 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 0);
+      assert.equal(process.stderr.write.callCount, 0);
     },
 
     'flow.signin.forgot-password': () => {
@@ -261,8 +259,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - forgot_pwd');
     },
 
@@ -275,8 +273,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_reg - have_account');
     },
 
@@ -289,8 +287,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - submit');
     },
 
@@ -303,8 +301,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - submit');
     },
 
@@ -317,8 +315,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_reg - submit');
     },
 
@@ -331,7 +329,7 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 0);
+      assert.equal(process.stderr.write.callCount, 0);
     },
 
     'screen.force-auth': () => {
@@ -343,8 +341,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - view');
     },
 
@@ -357,8 +355,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_login - view');
     },
 
@@ -371,8 +369,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_reg - view');
     },
 
@@ -385,8 +383,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_pref - view');
     },
 
@@ -408,8 +406,8 @@ define([
         utm_source: 'l',
         utm_term: 'm'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.deepEqual(arg, {
         app_version: APP_VERSION,
         device_id: 'b',
@@ -437,7 +435,7 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 0);
+      assert.equal(process.stderr.write.callCount, 0);
     },
 
     'settings.communication-preferences.optIn.success': () => {
@@ -449,8 +447,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_pref - newsletter');
       assert.equal(arg.user_properties.newsletter_state, 'subscribed');
     },
@@ -464,8 +462,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_pref - newsletter');
       assert.equal(arg.user_properties.newsletter_state, 'unsubscribed');
     },
@@ -479,7 +477,7 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 0);
+      assert.equal(process.stderr.write.callCount, 0);
     },
 
     'complete-reset-password.verification.success': () => {
@@ -500,8 +498,8 @@ define([
         utm_source: 'l',
         utm_term: 'm'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.deepEqual(arg, {
         app_version: APP_VERSION,
         device_id: 'b',
@@ -531,8 +529,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_email - click');
       assert.equal(arg.event_properties.email_type, 'login');
     },
@@ -546,8 +544,8 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 1);
-      const arg = log.info.args[0][0];
+      assert.equal(process.stderr.write.callCount, 1);
+      const arg = JSON.parse(process.stderr.write.args[0]);
       assert.equal(arg.event_type, 'fxa_email - click');
       assert.equal(arg.event_properties.email_type, 'registration');
     },
@@ -561,7 +559,7 @@ define([
         flowId: 'c',
         uid: 'd'
       });
-      assert.equal(log.info.callCount, 0);
+      assert.equal(process.stderr.write.callCount, 0);
     }
   });
 });